Small clean up in files

+++ Backend/src/main/java/routes/UserRoutes.java
@@ -2,6 +2,7 @@ package routes;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import express.Express;
+import express.http.Response;
 import models.Session;
 import models.User;
 import repositories.SessionRepository;
@@ -9,6 +10,11 @@ import repositories.UserRepository;
 import util.PasswordHash;
 
 import javax.servlet.http.Cookie;
+import java.net.HttpCookie;
+import java.sql.Date;
+import java.sql.Timestamp;
+import java.util.Calendar;
+import java.util.List;
 import java.util.Optional;
 
 
@@ -19,6 +25,9 @@ public class UserRoutes {
     private final UserRepository userRepository;
     private final SessionRepository sessionRepository;
 
+
+
+
     
     public UserRoutes(Express app, ObjectMapper mapper, UserRepository userRepository, SessionRepository sessionRepository) {
         this.app = app;
@@ -28,37 +37,49 @@ public class UserRoutes {
         this.userMethods();
     }
 
+    public Optional <Cookie> setupSession(String email, int id) {
+         Session session = new Session(email, new Timestamp(System.currentTimeMillis()), id);
+        if (sessionRepository.save(session).isPresent()) {
+            Cookie cookie = new Cookie("current-user", String.valueOf(session.getId()));
+            cookie.setMaxAge(3600);
+            cookie.setHttpOnly(true);
+            return Optional.of(cookie);
+
+         } else {
+        return  Optional.empty();
+    }
+    }
+
     public void userMethods() {
         app.post("/api/login-user", (req, res) -> {
-            Optional<User> user1 = userRepository.findByEmail(req.body().get("email").toString());
-            if (user1.isEmpty()) {
-                System.out.println("Rad 56: USER DOESNT EXIST!");
+            Optional<User> user = userRepository.findByEmail(req.body().get("email").toString());
+            if (user.isEmpty()) {
                 res.json("User doesnt exist!").status(401);
                 return;
             }
             
-            if (passwordHash.authenticate(user1.get().getPassword().toCharArray(), req.body().get("password").toString())) {
-                Session session = new Session(user1.get().getEmail());
-                if (sessionRepository.save(session).isPresent()) {
-                    System.out.println("Success!");
-                    req.session("current-user", session.getId());
-                    res.cookie("min-kaka", "hej").json(user1.get().getEmail()).status(201).redirect("/");
+            if (passwordHash.authenticate(user.get().getPassword().toCharArray(), req.body().get("password").toString())) {
+                Optional <Cookie> cookie = setupSession(user.get().getEmail(), user.get().getId());
+               // Session session = new Session(user.get().getEmail(), new Timestamp(System.currentTimeMillis()), user.get().getId());
+               // if (sessionRepository.save(session).isPresent()) {
+                 //   var today = Calendar.getInstance().getTime();;
+                   // var expiry = new Date(today.getTime() + 3600 * 1000);
+                //res.set("Set-Cookie", cookie);
+                if(cookie.isPresent()) {
+                    res.cookie(cookie.get()).json(user.get().getEmail()).status(201).redirect("/");
                 } else {
-                    System.out.println("Save failed.");
-                    res.json("Save failed").status(500);
+                    res.status(401).json("Sorry, could not log in..");
                 }
+                  //  res.set("Set-Cookie", "current-user=" + session.getId() + "; expires=" + expiry + "; HttpOnly; SameSite=strict");
 
             } else {
                 res.json("Wrong Password.").status(401);
             }
-
-
         });
         
         app.post("/api/register-user", (req, res) -> {
             Optional<User> userExists = userRepository.findByEmail(req.body().get("email").toString());
             if (userExists.isPresent()) {
-                System.out.println("User exists!");
                 res.json("User already exists.").status(401);
                 return;
             }
@@ -67,33 +88,29 @@ public class UserRoutes {
             user.setPassword(hashedPass);
             
             if (userRepository.save(user).isPresent()) {
-                System.out.println("Success!");
                 res.json("Success!").status(200);
             } else {
-                System.out.println("Save failed.");
                 res.json("Save failed").status(500);
             }
-            var userLoggedIn = mapper.writeValueAsString(user);
-            System.out.println("UNDER HÄR SPARAS SESSION: ");
-            req.session("current-user", userLoggedIn);
-            res.json(mapper.writeValueAsString(userLoggedIn));
+            res.status(201).redirect("/api/login-user");
         });
         
         app.get("/api/logout-user", (req, res) -> {
-            int session_id = Integer.parseInt(mapper.writeValueAsString(req.session("current-user")));
-            System.out.println("sezzion" + session_id);
+            String session_id = req.cookie("current-user");
+            Optional <Session> session = sessionRepository.findById(Integer.parseInt(session_id));
+            if(session.isPresent()) {
+                sessionRepository.deleteById(session.get());
+            }
 
-            Optional <Session> session = sessionRepository.findById(session_id);
-
-            sessionRepository.deleteById(session.get());
-            req.session("current-user", null);
-            res.json("Ok, logged out");
+            res.clearCookie("current-user", "/api").clearCookie("JSESSIONID", "/");
+            res.status(201).json("Successfully Logged out!");
         });
+
         
         app.get("/api/whoami", (req, res) -> {
-           // res.json(req.session("current-user"));
             res.json("hej");
         });
-        
     }
 }
+
+
